rectificatif sur remplacer, supprimer et chargerListeGrillesExistantes

+++ Codes/Modele/Jeu.rb
@@ -438,24 +438,45 @@ class Jeu
 
         if @profil != nil then
             
-            if @verbose then puts "suppression d'une partie!" end
+            if @verbose then 
+                
+                puts "suppression d'une partie!" 
+            end
+
             liste = Array.new()
             FileUtils.cd('Profil')
             FileUtils.cd(@profil.nom)
-            FileUtils.cd('Parties')
 
             if File.size('parties.yml') > 0 then
             
                 liste = YAML::load(File.open('parties.yml'))
-                sup = liste.delete(p)
-                File.delete('parties.yml')
-                File.open('parties.yml',"w"){|out| out.puts liste.to_yaml()}
-                FileUtils.cd('../../..')
-                return (sup == p)
+                listeNoms = Array.new()
+                
+                liste.each { |x|
+
+                    listeNoms.push(x[0])
+                }
+                if (index = listeNoms.index(unNom)) != nil then
+
+                    sup = liste.delete_at(index)
+                    File.delete('parties.yml')
+                    File.open('parties.yml',"w"){|out| out.puts liste.to_yaml()}
+                    FileUtils.cd('../..')
+                    return (sup == p)
+                else
+                   
+                    raise "erreur : aucune sauvegarde nommée #{unNom}"
+                end
             else
             
-               if @verbose then puts "Aucune partie n'est sauvegardée, impossible de supprimer" end
-                return false
+               if @verbose then 
+                   
+                   puts "Aucune partie n'est sauvegardée, impossible de supprimer" 
+               end
+               
+               FileUtils.cd('../..')
+
+               return false
 
             end
         else
@@ -529,9 +550,9 @@ class Jeu
 
         @partie = nil
     end
-#==============================================
+#===============================================
     #Gestion de la Partie /de l'éditeur en Cours
-#==============================================
+#===============================================
 
 # Méthode sauvegardant la partie en cours
 	def sauvegarderPartie(nomSauvegarde)
@@ -545,21 +566,26 @@ class Jeu
                 @partie.actualiser()
                 FileUtils.cd('Profil')
                 FileUtils.cd(@profil.nom)
-                if File.size("parties.yml") != 0 then
+
+                if File.size("parties.yml") > 0 then
                 
                     liste = YAML::load(File.open('parties.yml'))
                     listenom = Array.new()
                     liste.each { |x|
+
                         listenom.push(x[0])
                     }
                     if listenom.include?(nomSauvegarde) then
+                        
                         return false
                     end
                 end
+
                 liste.push([nomSauvegarde, @partie])
                 File.open('parties.yml',"w"){|out| out.puts liste.to_yaml()}
                 FileUtils.cd('../..')
                 return true
+
             else
 
                 @profil.ajouterUneGrille()
@@ -568,7 +594,7 @@ class Jeu
 
         else
             
-            raise "Aucune partie n'est en cours!"
+            raise "Rien (ni éditeur, ni partie) n'est en cours!"
         end
     end
